pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "RAG application using LangChain, Flask, and Pinecone"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain==0.3.26",
    "flask==3.1.1",
    "sentence-transformers==4.1.0",
    "pypdf==5.6.1",
    "python-dotenv==1.1.0",
    "langchain-pinecone==0.2.8",
    "langchain-openai==0.3.24",
    "langchain-community==0.3.26",
    "ipykernel>=6.30.0",
    "langchain-core>=0.3.72",
    "langchain-google-genai>=2.1.8",
    "huggingface-hub>=0.34.3",
    "pinecone>=5.0.0",
    "langchain-huggingface>=0.3.1",
]
